requirements.txt
Flask==1.1.2
numpy==1.20.3
Pillow==9.0.1
scikit_image==0.18.3
skimage==0.0
